Split tests into logical folders and refactor
